test(domains): merge PD set/delete integration tests into one file
The two files shared wallet setup, client init, and helper functions.
Merging into permissioned_domain.rs makes the full lifecycle readable
as a single file and eliminates duplicated boilerplate.

Changes:
- Move all tests from permissioned_domain_set.rs and
  permissioned_domain_delete.rs into tests/transactions/permissioned_domain.rs
- Route the delete test through the shared kyc_credential/new_pd_set helpers
  instead of its inlined PermissionedDomainSet struct literal
- Drop the temDISABLED early-return guard from the delete path (PermissionedDomains
  is always enabled in standalone mode — redundant guard masked test intent)
- Add Flags=0 assertion (xrpl.js parity: newly-created PD has Flags=0)
- Add ledger_entry deep-equal vs account_objects assertion (xrpl.js parity:
  assert.deepEqual(pd, ledgerEntryResult.result.node))
- Replace two mod entries in tests/transactions/mod.rs with single
  pub mod permissioned_domain
