+++ tests/transactions/mod.rs
@@ -33,8 +33,7 @@ pub mod payment_channel_claim;
 pub mod payment_channel_create;
 pub mod payment_channel_fund;
 pub mod payment_mpt;
-pub mod permissioned_domain_delete;
-pub mod permissioned_domain_set;
+pub mod permissioned_domain;
 pub mod set_regular_key;
 pub mod signer_list_set;
 pub mod submit_and_wait;
